requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
pyairtable==2.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
xmltodict==0.13.0
PyGithub==1.59.1
aiohttp==3.9.1
jinja2==3.1.2
python-multipart==0.0.6
latex==0.7.0
pypandoc==1.12
graphviz==0.20.1

